+++ app/Http/Controllers/SppController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\SppRequest;
 use App\Models\Spp;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 
@@ -18,7 +19,8 @@ class SppController extends Controller
         }
 
         return view('pages.admin.dataspp.index', [
-            'spp' => Spp::latest()->get(),
+            'spp' => Spp::latest()->orderBy('tahun', 'asc')->get(),
+            'user' => User::all()
         ]);
     }
 
@@ -58,7 +60,12 @@ class SppController extends Controller
 
     public function destroy(Spp $spp)
     {
-        $spp->delete();
-        return redirect(route('spp.index'))->with('info', 'Data berhasil dihapus!');
+        if (User::where('spp_id', $spp->id)->count() < 1) {
+            $spp->delete();
+            return redirect(route('spp.index'))->with('info', 'Data berhasil dihapus!');
+        } else {
+            return redirect(route('spp.index'))->withGagal('Data tidak dapat dihapus! Karena ada beberapa siswa dengan <b> SPP Tahun ' . $spp->tahun . '</b>!');
+        }
+       
     }
 }
